Add files via upload
This is the jupyter notebook file where I did most of my work for predicting pm2.5. You will be able to analyze all the EDA analysis for predicting AQI and then for predicting PM2.5. All of my feature store integration code is in this file. 
I only included the main code in my other python files which was working and created a workflow for them.