Add files via upload
This is the jupyter file of my all the code that I executed since the start of the project. This has the code for feature store integration (Hopsworks), Exploratory Data Analysis, model training and evaluaton.
All the final codes that were working; I included them in other files and generated a ci/cd pipeline on it. 
Dashboard is working on hugging face spaces and its ci/cd pipeline is on hugging face as well.